Rename "name" to "identifier"
