+++ docs/viewer/test.ts
@@ -1,10 +1,8 @@
 import { test } from "./test/util/cappucino";
 
-import * as CompositionTest from "./composition-test"
 import * as ComposeAlphaTest from "./compose-alpha-test"
 import * as ExampleFileTest from "./example-file-test"
 
-CompositionTest
 ComposeAlphaTest
 ExampleFileTest
