--- docs/viewer/test/fixtures/inherit_of_class/output.json
@@ -1,13 +1,0 @@
-{
-    "name": "",
-    "children": [
-        {
-            "attributes": {
-                "customdata:a": 2
-            },
-            "children": [],
-            "name": "/node1",
-            "type": "UsdGeom:Xform"
-        }
-    ]
-}
