--- docs/viewer/test/fixtures/inherit_of_def/output.json
@@ -1,28 +1,0 @@
-{
-  "children": [
-    {
-      "attributes": {
-        "customdata:a": 2
-      },
-      "children": [],
-      "name": "/node1",
-      "type": "UsdGeom:Xform"
-    },
-    {
-      "attributes": {
-        "customdata:a": 2
-      },
-      "children": [],
-      "name": "/node2",
-      "type": "UsdGeom:Xform"
-    },
-    {
-      "attributes": {
-        "customdata:a": 3
-      },
-      "name": "/node3",
-      "type": "UsdGeom:Xform"
-    }
-  ],
-  "name": ""
-}
