--- docs/viewer/test/fixtures/library_inherit/input_0.ifcx.json
@@ -1,29 +1,0 @@
-[
- {
-  "def": "class",
-  "type": "UsdGeom:Xform",
-  "name": "lib",
-  "children": [
-    {
-      "def": "def",
-      "type": "UsdGeom:Xform",
-      "name": "toilet",
-      "inherits": [
-        "</toiletclass>"
-      ]
-    }
-  ]
- },{
-  "def": "class",
-  "type": "UsdGeom:Xform",
-  "name": "toiletclass"
- },{
-  "def": "over",
-  "name": "toiletclass",
-  "attributes": {
-    "customdata": {
-      "a": 1
-    }
-  }
- }
-]
