--- docs/viewer/test/fixtures/library_inherit/input_1.ifcx.json
@@ -1,10 +1,0 @@
-[
-  {
-    "def": "def",
-    "type": "UsdGeom:Xform",
-    "name": "node1",
-    "inherits": [
-      "</lib/toilet>"
-    ]
-  }
-]
