--- docs/viewer/test/fixtures/library_inherit/output.json
@@ -1,13 +1,0 @@
-{
-    "children": [
-        {
-            "attributes": {
-                "customdata:a": 1
-            },
-            "children": [],
-            "name": "/node1",
-            "type": "UsdGeom:Xform"
-        }
-    ],
-    "name": ""
-}
