--- docs/viewer/test/fixtures/multiple_inherit/output.json
@@ -1,14 +1,0 @@
-{
-    "children": [
-        {
-            "attributes": {
-                "customdata:a": 1,
-                "customdata:b": 2
-            },
-            "children": [],
-            "name": "/node1",
-            "type": "UsdGeom:Xform"
-        }
-    ],
-    "name": ""
-}
