--- docs/viewer/test/fixtures/specific_override/input_0.ifcx.json
@@ -1,39 +1,0 @@
-[
-  {
-    "def": "def",
-    "type": "UsdGeom:Xform",
-    "name": "node1",
-    "children": [
-      {
-        "def": "def",
-        "name": "child1",
-        "inherits": [
-          "</child1class>"
-        ]
-      }
-    ]
-  },
-  {
-    "def": "class",
-    "type": "UsdGeom:Xform",
-    "name": "child1class"
-  },
-  {
-    "def": "over",
-    "name": "child1class",
-    "attributes": {
-      "customdata": {
-        "a": "classValue"
-      }
-    }
-  },
-  {
-    "def": "over",
-    "name": "node1/child1",
-    "attributes": {
-      "customdata": {
-        "a": "specificOverride"
-      }
-    }
-  }
-]
