--- docs/viewer/test/fixtures/specific_override/output.json
@@ -1,20 +1,0 @@
-{
-    "name": "",
-    "children": [
-        {
-            "name": "/node1",
-            "type": "UsdGeom:Xform",
-            "attributes": {},
-            "children": [
-                {
-                    "name": "/node1/child1",
-                    "type": "UsdGeom:Xform",
-                    "attributes": {
-                        "customdata:a": "specificOverride"
-                    },
-                    "children": []
-                }
-            ]
-        }
-    ]
-}
